+++ delphi-frontend/src/main/java/au/com/integradev/delphi/file/DefaultDelphiFileConfig.java
@@ -22,11 +22,10 @@ import au.com.integradev.delphi.preprocessor.DelphiPreprocessorFactory;
 import au.com.integradev.delphi.preprocessor.search.SearchPath;
 import java.nio.charset.Charset;
 import java.util.Set;
-import javax.annotation.Nullable;
 import org.sonar.plugins.communitydelphi.api.type.TypeFactory;
 
 public class DefaultDelphiFileConfig implements DelphiFileConfig {
-  private final String encoding;
+  private final Charset charset;
   private final Charset ansiCharset;
   private final DelphiPreprocessorFactory preprocessorFactory;
   private final TypeFactory typeFactory;
@@ -35,14 +34,14 @@ public class DefaultDelphiFileConfig implements DelphiFileConfig {
   private final boolean skipImplementation;
 
   DefaultDelphiFileConfig(
-      String encoding,
+      Charset charset,
       Charset ansiCharset,
       DelphiPreprocessorFactory preprocessorFactory,
       TypeFactory typeFactory,
       SearchPath searchPath,
       Set<String> definitions,
       boolean skipImplementation) {
-    this.encoding = encoding;
+    this.charset = charset;
     this.ansiCharset = ansiCharset;
     this.preprocessorFactory = preprocessorFactory;
     this.typeFactory = typeFactory;
@@ -51,10 +50,9 @@ public class DefaultDelphiFileConfig implements DelphiFileConfig {
     this.skipImplementation = skipImplementation;
   }
 
-  @Nullable
   @Override
-  public String getEncoding() {
-    return encoding;
+  public Charset getCharset() {
+    return charset;
   }
 
   @Override