+++ delphi-frontend/src/test/java/au/com/integradev/delphi/antlr/GrammarTest.java
@@ -275,7 +275,7 @@ class GrammarTest {
   void testUndefinedInaccessibleNestedIfDef() {
     fileConfig =
         DelphiFile.createConfig(
-            StandardCharsets.UTF_8.name(),
+            StandardCharsets.UTF_8,
             StandardCharsets.UTF_8,
             new DelphiPreprocessorFactory(
                 DelphiProperties.COMPILER_VERSION_DEFAULT, Platform.WINDOWS),